+++ Assets/Objects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 20d4224c929f4d52089f4968ff7bb11a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
