+++ Assets/Objects/Domai-Object-1_Building.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6e06c73897795b6e5a31f5937e7090f7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
