+++ Assets/Objects/Domai-Object-4_Flower-2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 69221c8db4bc96924ab991b03d1e0551
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
