+++ Assets/Objects/Domai-Object-5_Tree.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5bd4ce0d9856156d98379f2dd7b5b8b9
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
