+++ Assets/Objects/Garage_Linus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 981d0c251386457c7a6d797f3f3e9ae6
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
