+++ Assets/Objects/Lake_Linus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cfcc1f2a73968cc048e87d31ce45ed86
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
