+++ Assets/Objects/Street_Linus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6d7e02df1702fcfbd9608af0b4d9b3d0
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
