+++ Assets/Objects/Tree_Linus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e9e47b5ecd2934ee2b4e50ea29502d34
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
