+++ Assets/Objects/beehive_linus.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fb3fb7e4914cdfb889f96a4f69cccd5a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
