--- Assets/Player.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 00a1ff1ad9c72ff44bea97f1946d43f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
